refactor(runner): split idle reuse tests by behavior
