+++ crates/runner/src/cmd/start/tests/idle_reuse/mod.rs
@@ -1,0 +1,6 @@
+mod affinity;
+mod device_limits;
+mod drain;
+mod parking;
+mod status;
+mod workspace_cache;
